Handle client shutdown cleanly (#6)
Previously the `ReadError::Eof` error in `handle_stream` was unhandled, which meant that that same error occurred the next time an event occurred and was forwarded to the closed client stream (`forward_events`). This error led to the function returning, dropping `events_rx`. This in turn led to a panic the next time an event occurred , since the only receiver of `events_tx` had been dropped.